fix: modify home layout and delete outdated rollingpaper components
fix: modify home layout and delete past rollingpaper layout

fix: home layout, rollingpaper test layout deleted

fix: home layout, remove test layout

fix: home layout, rollingpaper test layout deleted

fix: home layout, rollingpaper test layout deleted

fix: modify home layout and delete rollingpaper outdated components

fix: modify home layout and delete rollingpaper outdated components
