+++ src/assets/index.ts
@@ -24,3 +24,4 @@ export { ReactComponent as ShareIcon2 } from './share.svg'
 export { ReactComponent as StickerBorder } from './sticker-border.svg'
 export { ReactComponent as StickerIcon } from './sticker-icn.svg'
 export { ReactComponent as ClipboardIcon } from './Text-files.svg'
+export { ReactComponent as Logo } from './logo.svg'
