+++ src/pages/Home/index.module.scss
@@ -5,20 +5,33 @@
   flex-direction: column;
   justify-content: space-between;
   align-items: center;
+  width: 100%;
   height: 100%;
-
+  padding: 42px 24px;
+  svg {
+    width: 100%;
+    height: auto;
+  }
+  .home-content {
+    display: flex;
+    flex-direction: column;
+    margin: 0 auto;
+    width: 100%;
+    padding: 0 35px;
+  }
   .card-container {
     display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    grid-template-rows: repeat(3, 1fr);
-    width: 70%;
+    grid-template-columns: repeat(3, minmax(85px, 1fr));
+    grid-template-rows: repeat(3, minmax(85px, 1fr));
+    width: 100%;
     aspect-ratio: 1 / 1;
-    @include txt.typograpy('basic', 14px, 400);
+    @include txt.typograpy('basic', 1rem, 400);
   }
 
   .home-description {
+    margin: 16px 0;
     text-align: center;
     line-height: 32px;
-    @include txt.typograpy('basic', 32px, 400);
+    @include txt.typograpy('basic', 24px, 400);
   }
 }
