+++ src/styles/layout/_layout.scss
@@ -10,11 +10,7 @@
   max-width: 480px;
   min-width: 380px;
   width: 100%;
-<<<<<<< HEAD
   height: 100vh;
-  padding: 32px;
-=======
->>>>>>> 005960a (feat: 롤링페이퍼 레이아웃 새버전)
   &.light {
     background-color: $mainBg;
     color: #000;
